refactor: inline workflow node resolution in factory
